+++ src/MauiControlsExtras/Controls/NumericUpDown.xaml.cs
@@ -875,6 +875,57 @@ public partial class NumericUpDown : TextStyledControlBase, IValidatable, Base.I
 
     #endregion
 
+    #region Mouse Wheel Support
+
+    /// <inheritdoc/>
+    protected override void OnHandlerChanged()
+    {
+        base.OnHandlerChanged();
+        SetupMouseWheelHandler();
+    }
+
+    private void SetupMouseWheelHandler()
+    {
+#if WINDOWS
+        if (Handler?.PlatformView is Microsoft.UI.Xaml.UIElement element)
+            element.PointerWheelChanged += OnWindowsPointerWheelChanged;
+#elif MACCATALYST
+        if (Handler?.PlatformView is UIKit.UIView view)
+        {
+            var panGesture = new UIKit.UIPanGestureRecognizer(OnMacScrollWheel);
+            panGesture.AllowedScrollTypesMask = UIKit.UIScrollTypeMask.Continuous | UIKit.UIScrollTypeMask.Discrete;
+            panGesture.MaximumNumberOfTouches = 0;
+            view.AddGestureRecognizer(panGesture);
+        }
+#endif
+    }
+
+#if WINDOWS
+    private void OnWindowsPointerWheelChanged(object sender, Microsoft.UI.Xaml.Input.PointerRoutedEventArgs e)
+    {
+        if (!IsFocused && _entry?.IsFocused != true) return;
+
+        var delta = e.GetCurrentPoint((Microsoft.UI.Xaml.UIElement)sender).Properties.MouseWheelDelta;
+        if (delta > 0) Increment();
+        else if (delta < 0) Decrement();
+        e.Handled = true;
+    }
+#endif
+
+#if MACCATALYST
+    private void OnMacScrollWheel(UIKit.UIPanGestureRecognizer recognizer)
+    {
+        if (!IsFocused && _entry?.IsFocused != true) return;
+
+        var translation = recognizer.TranslationInView(recognizer.View);
+        if (translation.Y > 2) Decrement();
+        else if (translation.Y < -2) Increment();
+        recognizer.SetTranslation(CoreGraphics.CGPoint.Empty, recognizer.View);
+    }
+#endif
+
+    #endregion
+
     #region Theme Updates
 
     /// <inheritdoc/>